Added frontend folder and ui

+++ backend/controllers/goalController.js
@@ -1,19 +1,21 @@
 const asyncHandler = require("express-async-handler");
 
 const Goal = require("../models/goalModel");
+const User = require("../models/userModel");
 const { text } = require("express");
 
 // @desc Get all goals
 // @route GET /api/goals
-// @access Public
+// @access Private
 const getGoals = asyncHandler(async (req, res) => {
-  const goals = await Goal.find({});
+  const goals = await Goal.find({ user: req.user.id });
+
   res.status(200).json(goals);
 });
 
 // @desc Set goals
 // @route POST /api/goals
-// @access Public
+// @access Private
 const setGoals = asyncHandler(async (req, res) => {
   if (!req.body.text) {
     res.status(400);
@@ -21,19 +23,32 @@ const setGoals = asyncHandler(async (req, res) => {
   }
   const goal = await Goal.create({
     text: req.body.text,
+    user: req.user.id,
   });
   res.status(200).json(goal);
 });
 
 // @desc Update goals
 // @route PUT /api/goals/:id
-// @access Public
+// @access Private
 const updateGoals = asyncHandler(async (req, res) => {
   const goal = await Goal.findById(req.params.id);
   if (!goal) {
     res.status(404);
     throw new Error("Goal not found");
   }
+
+  const user = await User.findById(req.user.id);
+  if (!user) {
+    res.status(401);
+    throw new Error("User not found");
+  }
+
+  if (goal.user.toString() !== user._id.toString()) {
+    res.status(401);
+    throw new Error("Not authorized");
+  }
+
   const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
     new: true,
   });
@@ -42,13 +57,23 @@ const updateGoals = asyncHandler(async (req, res) => {
 });
 // @desc Delete goals
 // @route DELETE /api/goals/:id
-// @access Public
+// @access Private
 
 const deleteGoals = asyncHandler(async (req, res) => {
   const goal = await Goal.findById(req.params.id);
   if (!goal) {
     res.status(404);
     throw new Error("Goal not found");
+  }
+  const user = await User.findById(req.user.id);
+  if (!user) {
+    res.status(401);
+    throw new Error("User not found");
+  }
+
+  if (goal.user.toString() !== user._id.toString()) {
+    res.status(401);
+    throw new Error("Not authorized");
   }
   // 替代 .remove()，使用 .findByIdAndDelete()
   await Goal.findByIdAndDelete(req.params.id);
